refactor: 사용자 계정 생성 이메일 전송 로직 개선 및 이메일 재전송 API 추가 완료 (#536)
