+++ build.gradle
@@ -34,7 +34,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation' // 유효성 검사
     implementation 'org.springframework.boot:spring-boot-starter-thymeleaf' // 템플릿 엔진
     implementation 'org.springframework.boot:spring-boot-starter-webflux' // WebFlux
-    developmentOnly 'org.springframework.boot:spring-boot-devtools' // 개발 도구
 
     // Security
     implementation 'org.springframework.boot:spring-boot-starter-security' // Spring Security
@@ -73,6 +72,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-mail' // Spring Mail
     implementation 'com.google.zxing:core:3.4.1' // QR 코드
     implementation 'com.google.zxing:javase:3.5.3' // QR 코드
+    implementation 'org.apache.commons:commons-lang3:3.17.0' // Apache Commons Lang
 
     // MapStruct
     implementation 'org.mapstruct:mapstruct:1.6.0'
